+++ docs/changes.rst
@@ -6,6 +6,8 @@ Changelog
 
 * :py:meth:`flask_views.base.TemplateResponseMixin.render_to_response` now
   accepts context data as ``context_data`` argument.
+* :py:class:`flask_views.json.JSONResponseMixin` and
+  :py:class:`flask_views.json.JSONView` classes added.
 
 
 0.1